reorganize files in folders
