+++ app/src/main/java/com/wearconnectivityexample/ui/screens/ChatDetailScreen.kt
@@ -1,4 +1,4 @@
-package com.example.wearconnectivityexample.ui
+package com.wearconnectivityexample.ui.screens
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.icons.Icons
