+++ app/src/main/java/com/wearconnectivityexample/ui/screens/ChatListScreen.kt
@@ -1,4 +1,4 @@
-package com.example.wearconnectivityexample.ui
+package com.wearconnectivityexample.ui.screens
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
